Add support for using two mark bits -- start/end
Marks the bits corresponding to the first and last word of the object
being marked.

Uses the bitmap to get the derive the size of the object. Note that
because the object may have been hashed, in which case it requires
an additional word to store the hashcode of the object. Because of this
requirement, we use an auxiliary bitmap positioned in the metadata
region just after the primary bitmap. The purpose of this bitmap is to
indicate whether the object has been hashed.