Fix bug caused by interleaving on mark bit flags
Initially the methods testAndSetLiveBits and testAndClearLiveBits were
dependent on both start and end bits of the object having been changed.
This update makes the return value of the function dependent only on
the state of the start marker -- with the end marker logically serving
more of the function of a size/extent marker than a liveness marker.

This fixes a bug that was caused by interleaving of threads when
setting or clearing the start/end bits.