Defer calculation of forwarded object size
Uses the newly exposed hash state methods to defer the decision of how
much space to allocate to forwarded (possibly moved) objects.

We have taken out the inclusion of moved size in getSizeFromBitmap()
method, and we decide how much to advance the toCursor by looking
at whether the object is actually going to be moved, as opposed to
just whether it has been hashed or not.